+++ src/pages/DashboardBuilder/builderHelpers.ts
@@ -1,9 +1,25 @@
 import { GridSize } from '@material-ui/core';
 
 export const GRAPH_TYPES = ['Chart', 'Table', 'Map'];
+
 export const CHART_TYPES = ['Line', 'Bar', 'Pie', 'Doughnut'];
-export const FILTERS_OPTIONS = ['Customer Account', 'Products', 'Market Segment', 'Sub Market Segment', 'Sales Rep', 'Product Category'];
+
+export const FILTERS_OPTIONS = [
+  { title: 'Customer Account', key: 'customerAccount' },
+  { title: 'Product', key: 'product' },
+  { title: 'Market Segment', key: 'marketSegment' },
+  { title: 'Sub Market Segment', key: 'subMarketSegment' },
+  { title: 'Sales Rep', key: 'salesRep' },
+  { title: 'Product Category', key: 'productCategory' },
+  { title: 'Country', key: 'country' }
+];
+
 export const baseURL = '/dashboard-master';
+
+export type IFilterType = {
+  title: string;
+  key: string;
+};
 
 export interface IFormDataType {
   uniqueId?: string;
@@ -14,7 +30,7 @@ export interface IFormDataType {
   hasFilters?: boolean;
   hasTableView?: boolean;
   hasExport?: boolean;
-  filters?: string[];
+  filters?: IFilterType[];
 }
 
 export const defaultFormConfigs = {